feat(finance): extract expenses list into feature components (Phase 6d)
- Extract Expenses list logic into ExpensesTable feature component
- Extract table JSX into dedicated ExpenseRow component
- Enforce keyboard accessibility (role="button", onKeyDown) for expense rows
