+++ app/users/dashboard/finance/expenses/page.tsx
@@ -1,137 +1,18 @@
 'use client';
 
-import { useState, useMemo } from 'react';
-import { useRouter } from 'next/navigation';
 import { useExpenses } from '@/hooks/expenses';
-import { Pagination, PageHeader } from '@/components/common';
+import { PageHeader } from '@/components/common';
 import { Button } from '@/components/shadcn/button';
-import { Badge } from '@/components/shadcn/badge';
-import { Checkbox } from '@/components/shadcn/checkbox';
-import { Card, CardContent, CardHeader } from '@/components/shadcn/card';
-import { Input } from '@/components/shadcn/input';
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from '@/components/shadcn/table';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/shadcn/select';
-import { DollarSign, CheckCircle, Clock, Search, Loader2 } from 'lucide-react';
-import {
-  Empty,
-  EmptyErrorMedia,
-  EmptyMedia,
-  EmptyHeader,
-  EmptyTitle,
-  EmptyDescription,
-} from '@/components/shadcn/empty';
+import { Card } from '@/components/shadcn/card';
+import { DollarSign, CheckCircle, Clock } from 'lucide-react';
 import Link from 'next/link';
 import { routes } from '@/nav';
-import {
-  ExpenseType,
-  ExpenseStatus,
-  expenseTypeLabels,
-  expenseStatusLabels,
-  expenseCategoryLabels,
-} from '@/types/finance/expense';
-
-const getStatusColor = (status: ExpenseStatus) => {
-  switch (status) {
-    case ExpenseStatus.paid:
-    case ExpenseStatus.reimbursed: {
-      return 'bg-green-100 text-green-800 dark:bg-green-900/20 dark:text-green-400';
-    }
-    case ExpenseStatus.approved: {
-      return 'bg-blue-100 text-blue-800 dark:bg-blue-900/20 dark:text-blue-400';
-    }
-    case ExpenseStatus.pending: {
-      return 'bg-yellow-100 text-yellow-800 dark:bg-yellow-900/20 dark:text-yellow-400';
-    }
-    case ExpenseStatus.draft: {
-      return 'bg-zinc-100 text-zinc-800 dark:bg-zinc-800 dark:text-zinc-400';
-    }
-    case ExpenseStatus.rejected: {
-      return 'bg-red-100 text-red-800 dark:bg-red-900/20 dark:text-red-400';
-    }
-    case ExpenseStatus.cancelled: {
-      return 'bg-zinc-100 text-zinc-800 dark:bg-zinc-800 dark:text-zinc-400';
-    }
-    default: {
-      return 'bg-zinc-100 text-zinc-800 dark:bg-zinc-800 dark:text-zinc-400';
-    }
-  }
-};
-
-const getTypeColor = (type: ExpenseType) => {
-  switch (type) {
-    case ExpenseType.direct: {
-      return 'bg-blue-100 text-blue-800 dark:bg-blue-900/20 dark:text-blue-400';
-    }
-    case ExpenseType.indirect: {
-      return 'bg-purple-100 text-purple-800 dark:bg-purple-900/20 dark:text-purple-400';
-    }
-    case ExpenseType.capital: {
-      return 'bg-green-100 text-green-800 dark:bg-green-900/20 dark:text-green-400';
-    }
-    case ExpenseType.operational: {
-      return 'bg-orange-100 text-orange-800 dark:bg-orange-900/20 dark:text-orange-400';
-    }
-    default: {
-      return 'bg-zinc-100 text-zinc-800 dark:bg-zinc-800 dark:text-zinc-400';
-    }
-  }
-};
+import { ExpenseStatus } from '@/types/finance/expense';
+import { ExpensesTable } from '@/features/expenses';
 
 export default function ExpensesPage() {
-  const router = useRouter();
   const { data: expenses = [], isLoading, isError } = useExpenses();
-  const [searchQuery, setSearchQuery] = useState('');
-  const [statusFilter, setStatusFilter] = useState<string>('all');
-  const [typeFilter, setTypeFilter] = useState<string>('all');
-  const [currentPage, setCurrentPage] = useState(1);
-  const [itemsPerPage, setItemsPerPage] = useState(10);
-  const [selectedIds, setSelectedIds] = useState<number[]>([]);
 
-  // Filter expenses based on search and filters
-  const filteredExpenses = useMemo(() => {
-    return expenses.filter((expense) => {
-      // Search filter
-      const searchLower = searchQuery.toLowerCase();
-      const matchesSearch =
-        !searchQuery ||
-        expense.expenseNumber.toLowerCase().includes(searchLower) ||
-        expense.description.toLowerCase().includes(searchLower) ||
-        expenseCategoryLabels[expense.category]
-          ?.toLowerCase()
-          .includes(searchLower);
-
-      // Status filter
-      const matchesStatus =
-        statusFilter === 'all' || expense.status === statusFilter;
-
-      // Type filter
-      const matchesType = typeFilter === 'all' || expense.type === typeFilter;
-
-      return matchesSearch && matchesStatus && matchesType;
-    });
-  }, [expenses, searchQuery, statusFilter, typeFilter]);
-
-  // Pagination
-  const startIndex = (currentPage - 1) * itemsPerPage;
-  const endIndex = startIndex + itemsPerPage;
-  const paginatedExpenses = filteredExpenses.slice(startIndex, endIndex);
-  const totalPages = Math.ceil(filteredExpenses.length / itemsPerPage);
-  const safePage = Math.min(currentPage, Math.max(1, totalPages));
-
-  // Calculate stats
   const totalExpenses = expenses.length;
   const paidExpenses = expenses.filter(
     (e) =>
@@ -142,38 +23,6 @@ export default function ExpensesPage() {
       e.status === ExpenseStatus.pending || e.status === ExpenseStatus.approved
   ).length;
   const totalAmount = expenses.reduce((sum, e) => sum + e.totalAmount, 0);
-
-  // Selection handlers
-  const handleSelectAll = (checked: boolean) => {
-    if (checked) {
-      setSelectedIds(paginatedExpenses.map((e) => e.id));
-    } else {
-      setSelectedIds([]);
-    }
-  };
-
-  const handleSelectOne = (id: number, checked: boolean) => {
-    if (checked) {
-      setSelectedIds([...selectedIds, id]);
-    } else {
-      setSelectedIds(selectedIds.filter((selectedId) => selectedId !== id));
-    }
-  };
-
-  const isAllSelected =
-    paginatedExpenses.length > 0 &&
-    selectedIds.length === paginatedExpenses.length;
-
-  const hasActiveFilters = Boolean(
-    searchQuery || statusFilter !== 'all' || typeFilter !== 'all'
-  );
-
-  const clearFilters = () => {
-    setSearchQuery('');
-    setStatusFilter('all');
-    setTypeFilter('all');
-    setCurrentPage(1);
-  };
 
   return (
     <div className="space-y-4 sm:space-y-6">
@@ -253,287 +102,11 @@ export default function ExpensesPage() {
         </div>
       </Card>
 
-      {/* Unified Card: search/filter toolbar + content + pagination */}
-      <Card>
-        <CardHeader className="flex flex-row flex-wrap items-center gap-3 border-b px-4 py-1">
-          {/* Search input */}
-          <div className="relative w-full max-w-xs">
-            <Search className="absolute top-1/2 left-3 size-3.5 -translate-y-1/2 text-zinc-400" />
-            <Input
-              value={searchQuery}
-              onChange={(e) => {
-                setSearchQuery(e.target.value);
-                setCurrentPage(1);
-              }}
-              placeholder="Search by expense number, description..."
-              className="h-8 pl-8 text-sm"
-            />
-          </div>
-          {/* Status filter */}
-          <Select
-            value={statusFilter}
-            onValueChange={(v) => {
-              setStatusFilter(v);
-              setCurrentPage(1);
-            }}
-          >
-            <SelectTrigger className="h-8 w-32 text-xs">
-              <SelectValue placeholder="Status" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="all">All Status</SelectItem>
-              <SelectItem value={ExpenseStatus.draft}>
-                {expenseStatusLabels[ExpenseStatus.draft]}
-              </SelectItem>
-              <SelectItem value={ExpenseStatus.pending}>
-                {expenseStatusLabels[ExpenseStatus.pending]}
-              </SelectItem>
-              <SelectItem value={ExpenseStatus.approved}>
-                {expenseStatusLabels[ExpenseStatus.approved]}
-              </SelectItem>
-              <SelectItem value={ExpenseStatus.paid}>
-                {expenseStatusLabels[ExpenseStatus.paid]}
-              </SelectItem>
-              <SelectItem value={ExpenseStatus.reimbursed}>
-                {expenseStatusLabels[ExpenseStatus.reimbursed]}
-              </SelectItem>
-              <SelectItem value={ExpenseStatus.rejected}>
-                {expenseStatusLabels[ExpenseStatus.rejected]}
-              </SelectItem>
-              <SelectItem value={ExpenseStatus.cancelled}>
-                {expenseStatusLabels[ExpenseStatus.cancelled]}
-              </SelectItem>
-            </SelectContent>
-          </Select>
-          {/* Type filter */}
-          <Select
-            value={typeFilter}
-            onValueChange={(v) => {
-              setTypeFilter(v);
-              setCurrentPage(1);
-            }}
-          >
-            <SelectTrigger className="h-8 w-32 text-xs">
-              <SelectValue placeholder="Type" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="all">All Types</SelectItem>
-              <SelectItem value={ExpenseType.direct}>
-                {expenseTypeLabels[ExpenseType.direct]}
-              </SelectItem>
-              <SelectItem value={ExpenseType.indirect}>
-                {expenseTypeLabels[ExpenseType.indirect]}
-              </SelectItem>
-              <SelectItem value={ExpenseType.capital}>
-                {expenseTypeLabels[ExpenseType.capital]}
-              </SelectItem>
-              <SelectItem value={ExpenseType.operational}>
-                {expenseTypeLabels[ExpenseType.operational]}
-              </SelectItem>
-            </SelectContent>
-          </Select>
-          {/* Rows per page — pushed to right */}
-          <div className="ml-auto flex items-center gap-2 border-l pl-3">
-            <span className="text-xs whitespace-nowrap text-zinc-500">
-              Rows per page
-            </span>
-            <Select
-              value={String(itemsPerPage)}
-              onValueChange={(v) => {
-                setItemsPerPage(Number(v));
-                setCurrentPage(1);
-              }}
-            >
-              <SelectTrigger className="h-8 w-16 text-xs">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                {[5, 10, 20, 50, 100].map((n) => (
-                  <SelectItem key={n} value={String(n)}>
-                    {n}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-        </CardHeader>
-
-        <CardContent className="p-0">
-          {(() => {
-            if (isLoading)
-              return (
-                <div className="flex justify-center py-12">
-                  <Loader2 className="h-8 w-8 animate-spin text-zinc-400" />
-                </div>
-              );
-            if (isError)
-              return (
-                <CardContent>
-                  <Empty variant="default">
-                    <EmptyErrorMedia>
-                      <DollarSign className="size-6" />
-                    </EmptyErrorMedia>
-                    <EmptyHeader>
-                      <EmptyTitle>Failed to load expenses</EmptyTitle>
-                      <EmptyDescription>
-                        An unexpected error occurred. Please try again.
-                      </EmptyDescription>
-                    </EmptyHeader>
-                  </Empty>
-                </CardContent>
-              );
-            if (paginatedExpenses.length > 0)
-              return (
-                <Table>
-                  <TableHeader>
-                    <TableRow>
-                      <TableHead className="w-12">
-                        <Checkbox
-                          checked={isAllSelected}
-                          onCheckedChange={handleSelectAll}
-                          aria-label="Select all"
-                        />
-                      </TableHead>
-                      <TableHead>Expense #</TableHead>
-                      <TableHead>Description</TableHead>
-                      <TableHead>Category</TableHead>
-                      <TableHead>Type</TableHead>
-                      <TableHead>Status</TableHead>
-                      <TableHead className="text-right">Amount</TableHead>
-                      <TableHead className="text-right">Paid</TableHead>
-                      <TableHead className="text-right">Balance</TableHead>
-                    </TableRow>
-                  </TableHeader>
-                  <TableBody>
-                    {paginatedExpenses.map((expense) => (
-                      <TableRow
-                        key={expense.id}
-                        className="hover:bg-muted/50 cursor-pointer"
-                        onClick={() =>
-                          router.push(
-                            routes.finance.expenses.detail(expense.id).href
-                          )
-                        }
-                      >
-                        <TableCell onClick={(e) => e.stopPropagation()}>
-                          <Checkbox
-                            checked={selectedIds.includes(expense.id)}
-                            onCheckedChange={(checked) =>
-                              handleSelectOne(expense.id, checked as boolean)
-                            }
-                            aria-label={`Select ${expense.expenseNumber}`}
-                          />
-                        </TableCell>
-                        <TableCell>
-                          <div className="flex items-center space-x-3">
-                            <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-linear-to-br from-blue-500 to-blue-600">
-                              <DollarSign className="h-5 w-5 text-white" />
-                            </div>
-                            <div>
-                              <p className="font-medium text-zinc-900 dark:text-zinc-100">
-                                {expense.expenseNumber}
-                              </p>
-                              {expense.billNumber && (
-                                <p className="text-xs text-zinc-500 dark:text-zinc-500">
-                                  Bill: {expense.billNumber}
-                                </p>
-                              )}
-                            </div>
-                          </div>
-                        </TableCell>
-                        <TableCell>
-                          <div className="max-w-xs truncate text-sm text-zinc-700 dark:text-zinc-300">
-                            {expense.description}
-                          </div>
-                        </TableCell>
-                        <TableCell>
-                          <span className="text-sm text-zinc-600 dark:text-zinc-400">
-                            {expenseCategoryLabels[expense.category] ||
-                              expense.category}
-                          </span>
-                        </TableCell>
-                        <TableCell>
-                          <Badge className={getTypeColor(expense.type)}>
-                            {expenseTypeLabels[expense.type]}
-                          </Badge>
-                        </TableCell>
-                        <TableCell>
-                          <Badge className={getStatusColor(expense.status)}>
-                            {expenseStatusLabels[expense.status]}
-                          </Badge>
-                        </TableCell>
-                        <TableCell className="text-right">
-                          <span className="font-semibold text-zinc-900 dark:text-zinc-100">
-                            ₹{expense.totalAmount.toLocaleString('en-IN')}
-                          </span>
-                        </TableCell>
-                        <TableCell className="text-right">
-                          <span
-                            className={
-                              expense.paidAmount > 0
-                                ? 'text-green-600 dark:text-green-400'
-                                : 'text-zinc-500'
-                            }
-                          >
-                            ₹{expense.paidAmount.toLocaleString('en-IN')}
-                          </span>
-                        </TableCell>
-                        <TableCell className="text-right">
-                          <span
-                            className={
-                              expense.balanceAmount > 0
-                                ? 'text-red-600 dark:text-red-400'
-                                : 'text-green-600 dark:text-green-400'
-                            }
-                          >
-                            ₹{expense.balanceAmount.toLocaleString('en-IN')}
-                          </span>
-                        </TableCell>
-                      </TableRow>
-                    ))}
-                  </TableBody>
-                </Table>
-              );
-            return (
-              <CardContent>
-                <Empty variant="default">
-                  <EmptyMedia variant="icon">
-                    <DollarSign className="size-6" />
-                  </EmptyMedia>
-                  <EmptyHeader>
-                    <EmptyTitle>No expenses found</EmptyTitle>
-                    <EmptyDescription>
-                      {hasActiveFilters
-                        ? 'Try adjusting your search or filters.'
-                        : 'Add your first expense to get started.'}
-                    </EmptyDescription>
-                  </EmptyHeader>
-                  {!hasActiveFilters && (
-                    <Button asChild>
-                      <Link href={routes.finance.expenses.new}>
-                        New Expense
-                      </Link>
-                    </Button>
-                  )}
-                </Empty>
-              </CardContent>
-            );
-          })()}
-        </CardContent>
-
-        <div className="flex items-center justify-between border-t px-4 py-2">
-          <span className="text-sm text-zinc-500">
-            {filteredExpenses.length === 0
-              ? '0 records'
-              : `${startIndex + 1}–${Math.min(endIndex, filteredExpenses.length)} of ${filteredExpenses.length} ${filteredExpenses.length === 1 ? 'expense' : 'expenses'}`}
-          </span>
-          <Pagination
-            currentPage={safePage}
-            totalPages={totalPages}
-            onPageChange={setCurrentPage}
-          />
-        </div>
-      </Card>
+      <ExpensesTable
+        expenses={expenses}
+        isLoading={isLoading}
+        isError={isError}
+      />
     </div>
   );
 }
