+++ features/expenses/index.ts
@@ -1,0 +1,2 @@
+export * from './components/expenses-table';
+export * from './components/expense-row';
